# Optimize Table Metadata calls for Iceberg tables