## Description
https://github.com/prestodb/presto/issues/21628

## Motivation and Context
Avoid redundant calls to getIcebergTable for a single query for the same table.

## Impact
Avoid redundant calls to getIcebergTable for a single query for the same table.

## Environment
Presto cluster - Local Mac
Data Size - tpch.tiny
Data source - S3

## Results
## Performance Improvement for TPCH Q4 - Tiny

```jsx
SELECT
  o.orderpriority,
  count(*) AS order_count
FROM
  iceberg.tpch.orders o
WHERE 
  o.orderdate >= DATE '1993-07-01'
  AND o.orderdate < DATE '1993-07-01' + INTERVAL '3' MONTH
  AND EXISTS (
    SELECT
      *
    FROM
      iceberg.tpch.lineitem l
    WHERE
      l.orderkey = o.orderkey
      AND l.commitdate < l.receiptdate
  )
GROUP BY
  o.orderpriority
ORDER BY
  o.orderpriority;
```

### Native Catalog

**Current Execution Time,**

![image](https://github.com/prestodb/presto/assets/7887476/d5156ab6-a4c8-4857-b6b4-ff9b20504fa2)


**After optimization,**

![image](https://github.com/prestodb/presto/assets/7887476/b7527f69-0ff4-4efc-a371-ac1a8e828758)


### Hive Catalog

**Current Execution Time,**

![image](https://github.com/prestodb/presto/assets/7887476/c4fc0178-bfe9-463f-a9b4-9678cd2d61d2)


**After optimization,**

![image](https://github.com/prestodb/presto/assets/7887476/e22d91e1-19ba-4073-a5fe-36190ca93353)

## JMH Benchmarks Results
**Environment:**
Presto cluster - Local Presto Query Runner
Data - Local

### Before
**Native Catalog,**
`Benchmark                                                  Mode  Cnt     Score     Error  Units
BenchmarkIcebergHadoopCatalog.benchmarkIcebergTableSelect  avgt   20   176.927 ±  76.529  ms/op
BenchmarkIcebergHadoopCatalog.benchmarkIcebergTableTpchQ4  avgt   20  2161.647 ± 127.355  ms/op`

**Hive Catalog,**
`Benchmark                                                Mode  Cnt     Score     Error  Units
BenchmarkIcebergHiveCatalog.benchmarkIcebergTableSelect  avgt   20   143.555 ±  50.956  ms/op
BenchmarkIcebergHiveCatalog.benchmarkIcebergTableTpchQ4  avgt   20  2147.876 ± 147.937  ms/op`

### After
**Native Catalog,**
`Benchmark                                                  Mode  Cnt     Score     Error  Units
BenchmarkIcebergHadoopCatalog.benchmarkIcebergTableSelect  avgt   20   118.736 ±  37.819  ms/op
BenchmarkIcebergHadoopCatalog.benchmarkIcebergTableTpchQ4  avgt   20  2038.915 ± 109.773  ms/op`

**Hive Catalog,**
`Benchmark                                                Mode  Cnt     Score    Error  Units
BenchmarkIcebergHiveCatalog.benchmarkIcebergTableSelect  avgt   20   138.459 ± 57.796  ms/op
BenchmarkIcebergHiveCatalog.benchmarkIcebergTableTpchQ4  avgt   20  2017.693 ± 96.666  ms/op`

**Note:**
In the benchmark, improvement numbers are not big since all the metadata is fetched from the local itself via QueryRunner. Earlier result screenshots are accessing tables from s3.

## Contributor checklist

- [ ] Please make sure your submission complies with our [development](https://github.com/prestodb/presto/wiki/Presto-Development-Guidelines#development), [formatting](https://github.com/prestodb/presto/wiki/Presto-Development-Guidelines#formatting), [commit message](https://github.com/prestodb/presto/wiki/Review-and-Commit-guidelines#commit-formatting-and-pull-requests), and [attribution guidelines](https://github.com/prestodb/presto/wiki/Review-and-Commit-guidelines#attribution).
- [ ] PR description addresses the issue accurately and concisely.  If the change is non-trivial, a GitHub Issue is referenced.
- [ ] Documented new properties (with its default value), SQL syntax, functions, or other functionality.
- [ ] If release notes are required, they follow the [release notes guidelines](https://github.com/prestodb/presto/wiki/Release-Notes-Guidelines).
- [ ] Adequate tests were added if applicable.
- [ ] CI passed.

## Release Notes

```
== RELEASE NOTES ==

Iceberg Changes
* Optimize Table Metadata calls for Iceberg tables
```