@jasonf20 2024-01-07
I think using a more permanent cache could help with subsequent queries to the same table, which is a common use case. I suppose that can be done in a separate PR, but for large table metadata objects, it can be significant.