@yingsu00 2024-01-25
> What are the benchmark results?

The benchmark has 2 queries, one simple select, one TPCH Q4. Reetika pasted the results above:

Environment:
Presto cluster - Local Presto Query Runner
Data - Local

Before
Native Catalog,
Benchmark                                                  Mode  Cnt     Score     Error  Units BenchmarkIcebergHadoopCatalog.benchmarkIcebergTableSelect  avgt   20   176.927 ±  76.529  ms/op BenchmarkIcebergHadoopCatalog.benchmarkIcebergTableTpchQ4  avgt   20  2161.647 ± 127.355  ms/op

Hive Catalog,
Benchmark                                                Mode  Cnt     Score     Error  Units BenchmarkIcebergHiveCatalog.benchmarkIcebergTableSelect  avgt   20   143.555 ±  50.956  ms/op BenchmarkIcebergHiveCatalog.benchmarkIcebergTableTpchQ4  avgt   20  2147.876 ± 147.937  ms/op

After
Native Catalog,
Benchmark                                                  Mode  Cnt     Score     Error  Units BenchmarkIcebergHadoopCatalog.benchmarkIcebergTableSelect  avgt   20   118.736 ±  37.819  ms/op BenchmarkIcebergHadoopCatalog.benchmarkIcebergTableTpchQ4  avgt   20  2038.915 ± 109.773  ms/op

Hive Catalog,
Benchmark                                                Mode  Cnt     Score    Error  Units BenchmarkIcebergHiveCatalog.benchmarkIcebergTableSelect  avgt   20   138.459 ± 57.796  ms/op BenchmarkIcebergHiveCatalog.benchmarkIcebergTableTpchQ4  avgt   20  2017.693 ± 96.666  ms/op

Note:
In the benchmark, improvement numbers are not big since all the metadata is fetched from the local itself via QueryRunner. Earlier result screenshots are accessing tables from s3.